refactor(core): Name the workflow reference registry and extractor files (no-changelog)
`references/index.ts` held two exports pulling in opposite directions — the
registry the importer applies bindings with, and the extractor the dependency
resolver collects requirements with — so no single name fitted it. Split them so
each file says what it holds, and each consumer imports only what it needs.
