+++ packages/cli/src/modules/n8n-packages/entities/workflow/references/workflow-references.ts
@@ -1,27 +1,12 @@
-import type { WorkflowEntity } from '@n8n/db';
-
 import { callerIdsReference } from './caller-ids.reference';
 import { errorWorkflowReference } from './error-workflow.reference';
 import { subWorkflowNodeReference } from './sub-workflow-node.reference';
 import type { EntityReference } from '../../reference';
 import type { WorkflowSubWorkflowRequirement } from '../workflow.types';
 
+/** Every place a workflow can reference another workflow, for both extract and rebind. */
 export const workflowReferences: Array<EntityReference<WorkflowSubWorkflowRequirement>> = [
 	subWorkflowNodeReference,
 	errorWorkflowReference,
 	callerIdsReference,
 ];
-
-export function extractWorkflowRequirements(
-	workflow: WorkflowEntity,
-): WorkflowSubWorkflowRequirement[] {
-	const byId = new Map<string, WorkflowSubWorkflowRequirement>();
-	for (const reference of workflowReferences) {
-		for (const requirement of reference.extract(workflow)) {
-			if (!byId.has(requirement.referencedWorkflowId)) {
-				byId.set(requirement.referencedWorkflowId, requirement);
-			}
-		}
-	}
-	return [...byId.values()];
-}
